// app/search/search-interface.ts
export interface ISearch {
  CaseId: string;
  Sid: string;
  FirstName: string;
  LastName: string;
  CaseRefNo: string;
  DateCreated: string;
  FormTypeId: string;
  TrackId: string;
  MasterStatus: string;
  ActiveStatus: string;
  IsActive: string;
  CustomsExciseCode: string;
  CompanyRegNo: string;
  TradingName: string;
  IncomeTaxNo: string;
  VATNo: string;

}